pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "openai>=1.72.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
